test(unit): converted the pacakge documentation tests to vitest
for #452
